+++ pkg/monitoring/alloy/monitoring-config.go
@@ -1,0 +1,13 @@
+package alloy
+
+type monitoringConfig struct {
+	Alloy monitoringConfigAlloy `json:"alloy"`
+}
+
+type monitoringConfigAlloy struct {
+	Controller monitoringConfigAlloyController `json:"controller"`
+}
+
+type monitoringConfigAlloyController struct {
+	Replicas int `json:"replicas"`
+}